C backend: generate code to buffer, simplify generator

+++ core/src/main/resources/org/teavm/backend/c/runtime.c
@@ -115,5 +115,5 @@ static int64_t currentTimeMillis() {
     struct timespec time;
     clock_gettime(CLOCK_REALTIME, &time);
 
-    return time.tv_sec * 1000 + round(time.tv_nsec / 1000000);
+    return time.tv_sec * 1000 + (int64_t) round(time.tv_nsec / 1000000);
 }
